Disable migration of databases
Since we turn right around and import a db backup this ends up
being wasted effort. The commands needed for the db migration
are still here, but commented out. If we move to installing
directly from the stable branch instead of a specific stable
tag that matches the version of the db backup then we should
restore the db migration steps.